New Feature: Metadata is now typed, support for directory properties (filename based on metadata, sequence number, default template) MetadataKey can be required and have order within a template.

+++ file-management/src/module/fileManagement/domain/task/CreateDefaultMetadataTemplates.java
@@ -1,8 +1,12 @@
 package module.fileManagement.domain.task;
 
 import module.fileManagement.domain.FileManagementSystem;
-import module.fileManagement.domain.MetadataKey;
-import module.fileManagement.domain.MetadataTemplate;
+import module.fileManagement.domain.metadata.DateTimeMetadata;
+import module.fileManagement.domain.metadata.IntegerMetadata;
+import module.fileManagement.domain.metadata.MetadataKey;
+import module.fileManagement.domain.metadata.MetadataTemplate;
+import module.fileManagement.domain.metadata.SeqNumberMetadata;
+import module.fileManagement.domain.metadata.StringMetadata;
 import pt.ist.fenixWebFramework.services.Service;
 
 public class CreateDefaultMetadataTemplates extends CreateDefaultMetadataTemplates_Base {
@@ -19,20 +23,35 @@ public class CreateDefaultMetadataTemplates extends CreateDefaultMetadataTemplat
 	return FileManagementSystem.getBundle().getString("label.task.create.default.metadata.templates");
     }
 
-    @Service
     public void resetTemplatesAndKeys() {
 	for (MetadataTemplate template : FileManagementSystem.getInstance().getMetadataTemplatesSet()) {
 	    for (MetadataKey key : template.getKeys()) {
-		key.delete();
+		if (!key.hasAnyMetadata()) {
+		    key.delete();
+		}
 	    }
 	    template.delete();
+	}
+    }
+
+    public void createDRHDeclarationTemplate() {
+	final String name = "Declaração DRH";
+	final MetadataTemplate metadataTemplate = FileManagementSystem.getInstance().getMetadataTemplate(name);
+	if (metadataTemplate == null) {
+	    MetadataTemplate template = new MetadataTemplate(name);
+	    template.addKey(MetadataKey.getOrCreateInstance("Nº Declaração", SeqNumberMetadata.class), 1, Boolean.TRUE);
+	    template.addKey(MetadataKey.getOrCreateInstance("Data", DateTimeMetadata.class), 2, Boolean.TRUE);
+	    template.addKey(MetadataKey.getOrCreateInstance("Destinatário", StringMetadata.class), 3, Boolean.TRUE);
+	    template.addKey(MetadataKey.getOrCreateInstance("Nº Mecanográfico", IntegerMetadata.class), 4, Boolean.TRUE);
+	    template.addKey(MetadataKey.getOrCreateInstance("Remetente", StringMetadata.class), 5, Boolean.TRUE);
+	    template.addKey(MetadataKey.getOrCreateInstance("Observações", StringMetadata.class), 6, Boolean.FALSE);
 	}
     }
 
     @Service
     public void executeTask() {
 	resetTemplatesAndKeys();
-	MetadataTemplate template = new MetadataTemplate("Documento");
+	createDRHDeclarationTemplate();
     }
 
 }
